requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pytesseract==0.3.10
Pillow==10.2.0
SQLAlchemy==2.0.28
pydantic==2.6.4
streamlit==1.32.0
requests==2.31.0
pandas==2.2.1
plotly==5.19.0
